Incorporated the SongStorage class into the Search class.
Added additional methods to the SongStorage class to handle common tasks like adding and displaying the table and to improve code reusability. 

Also added dates to the songs table for each song. An issue will arise in the future when a song becomes two days old where it's SongID will conflict with another song that released after it.